Modularize the custom command. All custom command should be placed in custom_comamnds_bot.py from now on
